requirements.txt
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
yt-dlp>=2023.3.4
pyshorteners>=1.0.1
python-slugify>=8.0.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.1
rich>=13.7.0
tiktok-uploader>=1.1.4